Remove overflow auto
This is the cause for the ember-power-select dropdown indent on the mandatendatabank homepage. Weirdly enough removing 'display: flex;' from 'au-c-main-header__title-group' also fixes this issue but causes the header to get messed up. Removing overflow:auto does not have any impact visuals and fixes the issue on all viewport sizes.