Register `<duet-date-picker>`  lazily
The initializer doesn't make the app wait which made it possible for `<AuDatePicker>` components to be rendered before `<duet-date-picker>` was registered. Since not all apps use the `<AuDatePicker>` component halting execution until it's loaded isn't the best idea.

This replaces the initializer with a "lazy-loaded" version in the component itself.

So now `<duet-date-picker>` will only be loaded if it's needed.