Add an option to localize the `AuDatePicker` component in Dutch
This adds a new config flag that replaces the English localization with a Dutch version when enabled. This option is disabled by default to keep backwards-compatibility, but it will become the default in the next major release since that's also the default language in all apps that use appuniversum. For this reason we also show a deprecation message so users know they need to manually switch to the Dutch version.