Remove the `@filter` argument
This was used in the initial implementation to provide the consumers with an alternative for ember-file-upload v4's validation logic. Now that we reimplemented that logic, we don't need to expose this hook. It could still be a nice feature, but we can add that in the future if there is a demand for it.