Add a route name where needed
Using an empty string as the route name would actually make the getter return `undefined` which seems to result in a href of `#`. Now that we removed the getter the empty string is forwarded directly to the LinkTo component which makes it output no `href` at all. This affects the styling.