AuDropdown refactoring
This removes the duplicated AuButton. We now use focus-trap's clickOutsideDeactivates event to detect when an outside click happened, and only call the close method if the target element is not the toggle button. This prevents the dropdown from instantly reopening again. The previous implementation worked around this by rendering a second button component which meant the click event was ignored.