Deprecate some functionality in the AuInput component
This deprecates the 2-way-binding and the masking feature in the AuInput component.

We were using `Input` internally, but this wasn't always usefull requiring some projects to create new components based on the native input instead. By deprecating the `Input` usage, all projects can now use `AuInput`, even for more advanced use-cases.

The masking functionality is now replaced by the `{{au-inputmask}}` modifier. Projects can apply the modifier instead of passing arguments. This makes the implementation a lot easier to maintain and projects that don't use the masking feature (and the AuDateInput component) won't ship the inputmask library if they are using Embroider "optimized".