Resolve "Unresolved dependencies" warnings
We're importing component from the package name space and Rollup didn't like that. We now added the package as an external and changed some import paths were needed. Converting everything to relative imports is counter intuitive and it's the main reason why we added the index.js file.

More info:  https://rollupjs.org/troubleshooting/#warning-treating-module-as-external-dependency