Fix the order of the "spacing" utilities
The responsive versions of the spacing utilities were generated before the standard ones. This causes issues if you use a "mobile first" pattern, where you start out with the non-responsive classes and add overrides with the responsive versions if needed. Since the responsive version was defined before the regular one, the regular one always wins. By changing the order it now works as expected.