Fix some whitespace issues
Some of our components generate inline(-*) elements which are affected by surrounding whitespace. We used a `{{~!~}}` comment for most of these cases, but it seems comments are stripped in production builds, even if the comment also has whitespace stripping enabled.

As a workaround we now remove the actual whitespace from the source code, even though it makes things harder to read.

Once we convert to a v2 addon we can strip the whitespace at build time.