Add a new `@onUpload` action to the `AuFileUpload` component
This is very similar to the `@onFinishUpload` action but with a more flexible signature that allows us to easily add new data in the future.

`@onUpload` also receives the `File` instance which allows consumers to display information about the selected file without having to fetch the corresponding data from the database first.